+++ widgets/src/Reflex/Dom/Modal/Class.hs
@@ -1,6 +1,6 @@
 {-| Description: Class for modal creators
 A class for widget that can have modals. A concrete implementation is in
-"Rhyolite.Frontend.Modal.Base".
+"Reflex.Dom.Modal.Base".
 -}
 
 {-# Language DefaultSignatures #-}
@@ -8,7 +8,7 @@ A class for widget that can have modals. A concrete implementation is in
 {-# Language MultiParamTypeClasses #-}
 {-# Language TypeFamilies #-}
 
-module Rhyolite.Frontend.Modal.Class where
+module Reflex.Dom.Modal.Class where
 
 import Control.Monad.Reader (MonadReader (ask), ReaderT (..))
 import Control.Monad.Trans (MonadTrans (lift))